+++ NEWS.md
@@ -3,11 +3,12 @@
 **Major breaking changes**
 
 This release overhauls the `tidy_table()` function of unpivotr to preserve the
-original data types of table cells.
+original data types of table cells and to support HTML tables.
 
 * `tidytable()` has been renamed `tidy_table()`.  `tidytable()` is an error,
     rather than a deprecation warning, because `tidy_table()` is so different
     from before.
+* There is a new `tidy_table()` method and vignette for HTML.
 * There is no `tidy_table()` method for matrices.  Convert matrices to
     data.frames first, choosing what to do with row and column names.
 * `tidy_table()` returns only relevant columns, according to the data types of
@@ -17,7 +18,9 @@ original data types of table cells.
     a column called `chr`.  The full list of column names is in `?tidy_table`
     and is `chr`, `cplx`, `cplx`, `dbl`, `fctr`, `int`, `lgl`, `list`.  The
     columns `fctr` and `list` are list-columns, where each element is itself a
-    list.  This means that factors with different levels are kept separate.
+    list.  This means that factors with different levels are kept separate.  For
+    HTML tables, an `html` column is returned containing the standalone HTML of
+    each cell.
 * Both `rowname` and `colname` arguments to `tidy_table()` now default to
     `FALSE`.
 * All functions consistently return a `tibble`.